+++ server/app.py
@@ -8,12 +8,12 @@ from flask import request
 from flask_restful import Resource
 
 # Add your model imports
-from models.users import User
-from models.followers import Follower
-from models.movies import Movie
-from models.shows import Show
-from models.movie_collections import MovieCollection
-from models.show_collections import ShowCollection
+from models.user import User
+from models.follower import Follower
+from models.movie import Movie
+from models.show import Show
+from models.movie_collection import MovieCollection
+from models.show_collection import ShowCollection
 
 # Views go here!
